Add highlight groups for tab modified flag
This adds support for highlights groups to the modified flag that can be
included in the tabline. The two highlight groups are TabModified and
TabModifiedSelected. TabModifiedSelected is the highlight group used for
the modified flag on the currently selected tab. TabModified is used for
all other tabs.

See setup below for sample usage.

```vim
    Plug '?/taboo.vim'                    " or however you install the plugin
    let g:taboo_modified_tab_flag = 'X'   " character shown for tabs w/
                                          " unsaved changes
    let g:taboo_format = '%N. %f%m'       " format for tabs
                                          " %N = tab number
                                          " %f = filename (tail)
                                          " %m = modified flag
                                          " see :help taboo for details

    " this diff adds support for these
    hi TabModified           ctermfg=137
    hi TabModifiedSelected   ctermfg=160
```